refactor moving things around
