+++ packages/leva/src/components/Number/Number.tsx
@@ -3,7 +3,7 @@ import { NumberInput } from '../ValueInput'
 import { LevaInputProps } from '../../types/'
 import { InternalNumberSettings } from './number-plugin'
 import { Label, Row } from '../UI'
-import { useDragNumber } from '../../utils/hooks'
+import { useDragNumber } from '../../hooks'
 import { RangeGrid, InnerNumberLabel } from './StyledNumber'
 import { RangeSlider } from './RangeSlider'
 import { useInputContext } from '../../context'
@@ -12,7 +12,7 @@ type NumberProps = LevaInputProps<number, InternalNumberSettings>
 
 export function Number({ label, displayValue, onUpdate, onChange, settings }: NumberProps) {
   const bind = useDragNumber({ settings, onDrag: onUpdate })
-  
+
   return (
     <NumberInput value={displayValue} onUpdate={onUpdate} onChange={onChange}>
       <InnerNumberLabel title={label.length > 1 ? label : ''} {...bind()}>
