+++ packages/leva/src/hooks/useInput.ts
@@ -1,7 +1,7 @@
 import { useCallback, useState, useEffect } from 'react'
 import shallow from 'zustand/shallow'
-import { useStoreContext } from '../../context'
-import { Data, DataItem } from '../../types'
+import { useStoreContext } from '../context'
+import { Data, DataItem } from '../types'
 
 const getInputAtPath = (data: Data, path: string) => {
   const { count, ...input } = data[path]
