+++ packages/leva/src/hooks/useVisiblePaths.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react'
-import { StoreType } from '../../store'
+import { StoreType } from '../store'
 import shallow from 'zustand/shallow'
 
 /**
