+++ packages/leva/src/useControls.ts
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useCallback, useState } from 'react'
 import { levaStore, StoreType } from './store'
 import { folder } from './helpers'
-import { useShallowMemo, useValuesForPath } from './utils/hooks'
+import { useShallowMemo, useValuesForPath } from './hooks'
 import { FolderSettings, Schema, SchemaToValues } from './types'
 import { useRenderRoot } from './components/Leva'
