+++ packages/leva/stories/components/decorator-reset.tsx
@@ -1,11 +1,11 @@
 import { StoryFnReactReturnType } from '@storybook/react/dist/client/preview/types'
 import * as React from 'react'
-import { store } from '../../src'
+import { levaStore } from '../../src'
 
 const DefaultStory = (Story: () => StoryFnReactReturnType) => {
   const [_, set] = React.useState(false)
   React.useEffect(() => {
-    store.dispose()
+    levaStore.dispose()
     set(true)
   }, [])
   return _ ? <Story /> : <></>
